+++ src/shared/types/index.ts
@@ -1,8 +1,2 @@
-/* eslint-disable @typescript-eslint/no-explicit-any */
-import React from 'react';
-
-export type ReactTagProps<
-  T extends
-    | keyof React.JSX.IntrinsicElements
-    | React.JSXElementConstructor<any>,
-> = React.ComponentProps<T>;
+export * from './tag-props';
+export * from './api';
