+++ src/shared/ui/select/ui/options/options.tsx
@@ -1,8 +1,7 @@
 import clsx from 'clsx';
 import React from 'react';
 
-// TODO: fix path
-import { Icon } from '../../../icon';
+import { Icon } from '@/shared/ui';
 
 import { Option } from '../../config';
 
@@ -32,8 +31,8 @@ export const SelectOptions: React.FC<SelectOptionsProps> = ({
           key={option.value}
           className={clsx(
             styles.option,
-            (isSelected || (nothingSelected && option.value === ''))
-              && styles.selected,
+            (isSelected || (nothingSelected && option.value === '')) &&
+              styles.selected,
           )}
         >
           {option.label}
